Extract rendering logic to Output formatters
- FindOutputFormatter: all rendering from FindCommand (Markout tables,
  oneline, name-only). Deduplicates row-building that was repeated in
  WriteMultiPatternOutput and WriteMarkoutOutput.
- DiffOutputFormatter: all rendering from DiffCommand (stat, full
  markdown, name-only, section writing, summary counts).
- Commands now only contain orchestration + data logic.
- Both commands no longer import Markout directly.
- All 385 tests pass (263 app + 122 services)
